seperated test concerns
